% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstallr.R
\name{find_used_packages}
\alias{find_used_packages}
\title{Find Files Where Specific Packages Are Used}
\usage{
find_used_packages(packages, path = NULL, pattern = NULL)
}
\arguments{
\item{packages}{Vector of packages to look for}

\item{path}{Directory which is scanned recursively. Default is the working directory.}

\item{pattern}{Regex to identify R source files. Default is .*\\.(R|r|Rnw|Rhtml|Rpres|Rmd)$}
}
\value{
A data.frame with the files which are using the specified packages
}
\description{
Find Files Where Specific Packages Are Used
}
\examples{
\dontrun{
find_used_packages('dplyr', path = '../')
}
}

