% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstallr.R
\name{missing_packages}
\alias{missing_packages}
\title{Title}
\usage{
missing_packages(packages, ...)
}
\arguments{
\item{packages}{Vector of package names}

\item{...}{Parameters passed to available.packages()}
}
\value{
Vector of not installed packages
}
\description{
Title
}

