% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstallr.R
\name{show_missing_packages}
\alias{show_missing_packages}
\title{Show used but not installed packages}
\usage{
show_missing_packages(path = NULL, pattern = NULL, ...)
}
\arguments{
\item{path}{Directory which is scanned recursively. Default is the working directory.}

\item{pattern}{Regex to identify R source files. Default is .*\\.(R|r|Rnw|Rhtml|Rpres|Rmd)$}

\item{...}{Parameters passed to available.packages()}
}
\value{
A data.frame with missing packages
}
\description{
Show used but not installed packages
}
\details{
\code{show_missing_packages()} searches missing packages and checks if they are available on CRAN
}
\examples{
\dontrun{
show_missing_packages('../')
}
}

