% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstallr.R
\name{show_package_stats}
\alias{show_package_stats}
\title{Show Used Packages}
\usage{
show_package_stats(path = NULL, pattern = NULL)
}
\arguments{
\item{path}{Directory which is scanned recursively. Default is the working directory.}

\item{pattern}{Regex to identify R source files. Default is .*\\.(R|r|Rnw|Rhtml|Rpres|Rmd)$}
}
\value{
A aggregated data.frame
}
\description{
Show Used Packages
}
\examples{
show_package_stats('../')
}

