% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstallr.R
\name{reinstallr}
\alias{reinstallr}
\title{reinstallr}
\usage{
reinstallr(path = NULL, pattern = NULL, ...)
}
\value{
Installs packages or if not interactive a list of missing packages
}
\description{
reinstallr
}

